# packages/feed/openapi.yaml
openapi: 3.1.0
info:
  title: BarlesChailey Dispatch Feed
  version: "1.0.0"
  description: |
    Standardized MCFRS dispatch feed and responder alerting API.

    **Authentication**
    * Members: sign in with Neon Auth, fetch a short-lived Neon JWT (`/token`), and exchange it once at
      `POST /v1/auth/exchange` for a 30-day API token. Use `Authorization: Bearer <API token>` everywhere else.
    * Servers / dashboards: `X-Api-Key: bck_...` with scopes `read:feed` and/or `ingest`.
    * Station ingest: `X-Push-Secret: <shared secret>` (legacy-compatible with sta03-dashboard `/push`).
servers:
  - url: https://barleschailey-api.<account>.workers.dev
  - url: https://barleschailey-api-staging.<account>.workers.dev
security:
  - bearer: []
  - apiKey: []
...
paths:
...
  /v1/me/preferences:
    put:
      tags: [me]
      summary: Update alert preferences (partial)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AlertPreferences" } } } }
      responses: { "200": { description: OK }, "403": { description: pending approval } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKey: { type: apiKey, in: header, name: X-Api-Key }
    pushSecret: { type: apiKey, in: header, name: X-Push-Secret }
  schemas:
    IncidentCategory: { type: string, enum: [fire, rescue, ems, hazmat, service, other] }
    Severity: { type: string, enum: [low, normal, high, critical] }
...
    AlertPreferences:
      type: object
      properties:
        enabled: { type: boolean }
        radiusMiles: { type: number, minimum: 0.25, maximum: 50 }
        categories: { type: array, items: { $ref: "#/components/schemas/IncidentCategory" }, description: Empty means all }
        minSeverity: { $ref: "#/components/schemas/Severity" }
        locationMaxAgeHours: { type: integer, minimum: 1, maximum: 168 }
        home: { type: [object, "null"], properties: { lat: { type: number }, lon: { type: number }, radiusMiles: { type: number }, label: { type: [string, "null"] } } }
        quietHours: { type: [object, "null"], properties: { start: { type: string, example: "23:00" }, end: { type: string, example: "06:30" }, allowCritical: { type: boolean } } }
        snoozeUntil: { type: [string, "null"], format: date-time }
        alertOnUpgrade: { type: boolean }
...